refactor: cleanup cypress tools
